FIx input of eXamine_interactions
Depending on the name of the network (here HPRD), you have an irregular filename in the networks subdirectory. The mapping between the name and the filename is defined in the config file. So, to determine the input for this rule, we have to look this up in the config. This has to be done **after** the parsing, when jobs are determined and wildcard values are known. Before, Snakemake only knows about rules, not about jobs and wildcards. Hence, the execution of this lookup has to be deferred to later (the so-called DAG phase, see [here](http://snakemake.readthedocs.io/en/stable/tutorial/advanced.html#step-3-input-functions)).